View\Controller
The old abstract class Controller became View becouse it contains the code which helps with the Scene in MenuView.
